+++ R-Type/server/includes/Network/PlayerCommunicationManager.hpp
@@ -48,10 +48,10 @@ class PlayerCommunicationManager : public NoCopyable, public PlayerPacketBuilder
 
 	// send commands
 	public:
-        void	sendMoveResource(const Peer &peer, uint64_t id, IResource::Type type, short x, short y, short angle);
+        void	sendMoveResource(const Peer &peer, uint64_t id, IResource::Type type, float x, float y, float angle);
 		void	sendDestroyResource(const Peer &peer, uint64_t id);
 		void	sendUpdateScore(const Peer &peer, uint64_t id, const std::string &pseudo, uint64_t score);
-		void	sendTimeElapsedPing(const Peer &peer, double timeElapsed);
+        void	sendTimeElapsedPing(const Peer &peer, float timeElapsed);
 
 	// build cmd
 	private:
