+++ R-Type/server/includes/Script/IScriptCommand.hpp
@@ -10,8 +10,8 @@ class IScriptCommand {
         virtual ~IScriptCommand() = default;
     public:
         virtual IScriptCommand* clone(void) const = 0;
-        virtual double getFrame(void) const = 0;
-        virtual void setFrame(double) = 0;
+        virtual float getFrame(void) const = 0;
+        virtual void setFrame(float) = 0;
         virtual IScriptCommand::Instruction getInstruction(void) const = 0;
         virtual void setInstruction(IScriptCommand::Instruction) = 0;
 };
@@ -19,8 +19,8 @@ class IScriptCommand {
 class ScriptSpawn : public IScriptCommand {
     public:
         virtual IScriptCommand* clone(void) const { return new ScriptSpawn(*this); }
-        virtual double getFrame(void) const { return mFrame; }
-        virtual void setFrame(double frame) { mFrame = frame; }
+        virtual float getFrame(void) const { return mFrame; }
+        virtual void setFrame(float frame) { mFrame = frame; }
         virtual IScriptCommand::Instruction getInstruction(void) const { return mInstruction; }
         virtual void setInstruction(IScriptCommand::Instruction instruction) { mInstruction = instruction; }
     public:
@@ -47,18 +47,18 @@ class ScriptSpawn : public IScriptCommand {
     public:
         const std::string& getSpawnName(void) const { return mSpawnName; }
         void setSpawnName(const std::string& spawnName) { mSpawnName = spawnName; }
-        short getX(void) const { return mX; }
-        void setX(short x) { mX = x; }
-        short getY(void) const { return mY; }
-        void setY(short y) { mY = y; }
-        short getAngle(void) const { return mAngle; }
-        void setAngle(short angle) { mAngle = angle; }
+        float getX(void) const { return mX; }
+        void setX(float x) { mX = x; }
+        float getY(void) const { return mY; }
+        void setY(float y) { mY = y; }
+        float getAngle(void) const { return mAngle; }
+        void setAngle(float angle) { mAngle = angle; }
     private:
-        double mFrame;
+        float mFrame;
         IScriptCommand::Instruction mInstruction;
     private:
         std::string mSpawnName;
-        short mX;
-        short mY;
-        short mAngle;
+        float mX;
+        float mY;
+        float mAngle;
 };
